Do not sys.exit(2) when device is not found, retry instead
If broadlink-mqtt starts to run before the broadlink device is up, it
fails and it doesn't reconnect.
This happens only in the first connection. Once connected,
broadlink-mqtt handles disconection well.

This behaviour is annoying when configuring broadlink-mqtt as a systemd
service, since it will get an error and won't recover.

So instead of exiting, just retry, letting the user to turn on the
broadlink device, or turn off some devices in case multiple broadlinks
were found, and the system will auto-connect without user intervention.